+++ managed_components/esphome__esp-audio-libs/src/quantization_utils.cpp
@@ -1,5 +1,7 @@
 #include "quantization_utils.h"
 
+#include <cstddef>
+
 namespace esp_audio_libs {
 namespace quantization_utils {