+++ src/esphome/components/dashboard_import/dashboard_import.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace esphome::dashboard_import {
+
+const char *get_package_import_url();
+void set_package_import_url(const char *url);
+
+}  // namespace esphome::dashboard_import