+++ src/esphome/core/build_info_data.cpp
@@ -1,0 +1,15 @@
+// Auto-generated build_info data
+#include "esphome/core/build_info_data.h"
+
+namespace esphome {
+const uint32_t ESPHOME_CONFIG_HASH = 0x409d2e16U;  // NOLINT
+const time_t ESPHOME_BUILD_TIME = 1779386961;  // NOLINT
+const size_t ESPHOME_COMMENT_SIZE = 1;  // NOLINT
+#ifdef USE_ESP8266
+const char ESPHOME_BUILD_TIME_STR[] PROGMEM = "2026-05-21 19:09:21 +0100";
+const char ESPHOME_COMMENT_STR[] PROGMEM = "";
+#else
+const char ESPHOME_BUILD_TIME_STR[] = "2026-05-21 19:09:21 +0100";
+const char ESPHOME_COMMENT_STR[] = "";
+#endif
+}  // namespace esphome