+++ src/esphome/core/build_info_data.h
@@ -1,13 +1,21 @@
 #pragma once
-// Auto-generated build_info data
-#define ESPHOME_CONFIG_HASH 0x02cff806U  // NOLINT
-#define ESPHOME_BUILD_TIME 1778132398  // NOLINT
-#define ESPHOME_COMMENT_SIZE 1  // NOLINT
+// Auto-generated build_info declarations
+#include <cstddef>
+#include <cstdint>
+#include <ctime>
 #ifdef USE_ESP8266
 #include <pgmspace.h>
-static const char ESPHOME_BUILD_TIME_STR[] PROGMEM = "2026-05-07 06:39:58 +0100";
-static const char ESPHOME_COMMENT_STR[] PROGMEM = "";
-#else
-static const char ESPHOME_BUILD_TIME_STR[] = "2026-05-07 06:39:58 +0100";
-static const char ESPHOME_COMMENT_STR[] = "";
 #endif
+
+namespace esphome {
+extern const uint32_t ESPHOME_CONFIG_HASH;
+extern const time_t ESPHOME_BUILD_TIME;
+extern const size_t ESPHOME_COMMENT_SIZE;
+#ifdef USE_ESP8266
+extern const char ESPHOME_BUILD_TIME_STR[] PROGMEM;
+extern const char ESPHOME_COMMENT_STR[] PROGMEM;
+#else
+extern const char ESPHOME_BUILD_TIME_STR[];
+extern const char ESPHOME_COMMENT_STR[];
+#endif
+}  // namespace esphome